@@ -3,7 +3,7 @@ name = "arithmetic-parser-cli"
 version = "0.3.0"
 authors = ["Alex Ostrovski <ostrovski.alex@gmail.com>"]
 edition = "2021"
-rust-version = "1.57"
+rust-version = "1.59"
 readme = "README.md"
 license = "MIT OR Apache-2.0"
 description = "CLI / REPL for arithmetic expressions."
