@@ -15,7 +15,7 @@ impl MockTypes for MockedTypesList {
 type Grammar = WithMockedTypes<F64Grammar, MockedTypesList>;
 
 fn read_file(path: &str) -> String {
-    fs::read_to_string(path).unwrap_or_else(|err| panic!("Cannot read file {}: {}", path, err))
+    fs::read_to_string(path).unwrap_or_else(|err| panic!("Cannot read file {path}: {err}"))
 }
 
 fn check_sample(code_sample: &str) {
