@@ -99,7 +99,7 @@ fn methods_and_multiple_unary_ops() {
     assert_eq!(*expr.fragment(), "--1.abs()");
     let inner_expr = match expr.extra {
         Expr::Unary { inner, .. } => inner,
-        _ => panic!("Unexpected expr: {:?}", expr),
+        _ => panic!("Unexpected expr: {expr:?}"),
     };
     assert_eq!(*inner_expr.fragment(), "-1.abs()");
     assert_matches!(inner_expr.extra, Expr::Unary { .. });
@@ -142,7 +142,7 @@ fn chained_comparisons() {
         let err = binary_expr::<FieldGrammar, Complete>(InputSpan::new(input)).unwrap_err();
         let err = match err {
             NomErr::Failure(err) => err,
-            _ => panic!("Unexpected error: {:?}", err),
+            _ => panic!("Unexpected error: {err:?}"),
         };
         assert_matches!(err.kind(), ErrorKind::ChainedComparison);
         assert_eq!(*err.span().fragment(), input);
@@ -155,7 +155,7 @@ fn chained_comparisons_with_larger_context() {
     let err = binary_expr::<FieldGrammar, Complete>(InputSpan::new(input)).unwrap_err();
     let err = match err {
         NomErr::Failure(err) => err,
-        _ => panic!("Unexpected error: {:?}", err),
+        _ => panic!("Unexpected error: {err:?}"),
     };
     assert_matches!(err.kind(), ErrorKind::ChainedComparison);
     assert_eq!(*err.span().fragment(), "1 + 2 > x.abs() == 3");
